vault-service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-jose[cryptodigest]
passlib[bcrypt]
python-multipart
httpx
pycryptodome